requirements.txt
faiss-cpu
open3d
torch-scatter
opencv-python
pycocotools
matplotlib
onnxruntime
onnx
git+https://github.com/facebookresearch/segment-anything.git
numpy==1.26.1
Pillow
huggingface_hub
einops
safetensors
# torch==2.4.1 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
# Set FORCE_CUDA=1 for pytorch3d 
"git+https://github.com/facebookresearch/pytorch3d.git"
# Set FORCE_CUDA=1 for pytorch3d 
scikit-image
pyyaml  # For configuration file parsing